+++ apps/kyb-app/src/common/consts/consts.ts
@@ -1,0 +1,1 @@
+export const ARRAY_VALUE_INDEX_PLACEHOLDER = `{INDEX}`;