+++ apps/kyb-app/src/components/atoms/Stepper/components/atoms/Breadcrumbs/helpers/types.ts
@@ -1,0 +1,10 @@
+import {
+  BreadcrumbState,
+  BreadcrumbTheme,
+} from '@/components/atoms/Stepper/components/atoms/Breadcrumbs/types';
+
+export type ElementPropsPicker<TElementProps> = (
+  state: BreadcrumbState,
+  active: boolean,
+  theme: BreadcrumbTheme,
+) => TElementProps;